+++ src/universal/config/index.ts
@@ -1,4 +1,4 @@
-export * from './app';
+export * from './feature-toggles';
 export * from './thema';
 export * from './env';
 export * from './myarea-datasets';
